// src/core/interfaces/services/common.ts
export interface XXSAE_INV_DISPONIBLE_OS {
    ORGANIZATION_ID: number;
    ORGANIZATION_CODE:string;
    INVENTORY_ITEM_ID: number;
    ITEM:string;
    NO_INVENTARIO:string;
    NO_BIEN_SIAB: number;
    NO_GESTION:string;
    SUBINVENTORY_CODE:string;
    LOCATOR_ID: number;
    LOCATOR:string;
    UOM_CODE:string;
    DESCRIPTION:string;
    DISPONIBLE: number;
    RESERVADO: number;
    ROW_NUM: number;
  }